Upload, Download, Payment support
Now you can upload files and you will receive an uuid to access that file.
Specify your bitcoin address and the amount of bitcoins you want to get
for each download.

On the download page you need to enter the uuid code to request a
bitcoin address for payment.  If you sent a transaction to that address
you can access the download link.
When downloading, your transaction will be sent to the uploader.